Created the most simple bindings for java.
These bindings need a lot of work to improve quality but now are working decently.

It is important to remark that you need libdbus-java. In addition you need to include the /usr/share/java/dbus.jar and particularly the /usr/lib/jni/libunix-java.so jni library. It needs to be included in the eclipse configuration also.